last commit

+++ SpringBootEmployeeCRUD/src/main/java/com/example/demo/dao/EmployeeDaoI.java
@@ -6,8 +6,8 @@ import com.example.demo.entities.Employee;
 public interface EmployeeDaoI {
 	
 	Employee CreateEmployee(Employee emp);
-	Employee findEmployeeById(int empId);
+	Employee findEmployeeById(long empId);
 	Employee updateEmployee(Employee emp); 
 	List<Employee> findAllEmployees();
-	void deleteEmployee(int empId);
+	void deleteEmployee(long empId);
 }
